% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_plot.R
\name{demotrace_path}
\alias{demotrace_path}
\title{example nvidia-smi trace dataset}
\usage{
demotrace_path()
}
\value{
data.frame
}
\description{
example nvidia-smi trace dataset
}
\examples{
plot_trace(demotrace_path())
}
