% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_plot.R
\name{plot_trace}
\alias{plot_trace}
\title{visualize various aspects of performance with GPU}
\usage{
plot_trace(pa)
}
\arguments{
\item{pa}{path to an nvidia-smi trace file}
}
\value{
plot
}
\description{
visualize various aspects of performance with GPU
}
\examples{
plot_trace(demotrace_path())
}
