% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lkseq.R
\name{view_sequences}
\alias{view_sequences}
\title{viewer for event-sequence data}
\usage{
view_sequences(nrec = 5)
}
\arguments{
\item{nrec}{numeric defaults to 5}
}
\description{
viewer for event-sequence data
}
\examples{
tt = view_sequences()
strsplit(tt[[1]][[2]], " ")
}
